+++ frontend/src/components/BudgetLineItems/AllBLIRow/AllBLIRow.js
@@ -4,7 +4,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faClock } from "@fortawesome/free-regular-svg-icons";
 import CurrencyFormat from "react-currency-format";
 import TableTag from "../../UI/TableTag";
-import ChangeIcons from "../../UI/ChangeIcons";
+import ChangeIcons from "../ChangeIcons";
 import TableRowExpandable from "../../UI/TableRowExpandable";
 import {
     formatDateNeeded,
